+++ imagepipeline/src/main/jni/webpimage/webp.cpp
@@ -227,7 +227,7 @@ jobject WebPImage_nativeCreateFromByteVector(JNIEnv* pEnv, std::vector<uint8_t>&
   jobject ret = pEnv->NewObject(
       sClazzWebPImage,
       sWebPImageConstructor,
-      (jint) spNativeContext.get());
+      (long) spNativeContext.get());
   if (ret != nullptr) {
     // Ownership was transferred.
     spNativeContext->refCount = 1;
@@ -279,7 +279,7 @@ std::unique_ptr<WebPImageNativeContext, WebPImageNativeContextReleaser>
   std::unique_ptr<WebPImageNativeContext, WebPImageNativeContextReleaser> ret(nullptr, releaser);
   pEnv->MonitorEnter(thiz);
   WebPImageNativeContext* pNativeContext =
-      (WebPImageNativeContext*) pEnv->GetIntField(thiz, sWebPImageFieldNativeContext);
+      (WebPImageNativeContext*) pEnv->GetLongField(thiz, sWebPImageFieldNativeContext);
   if (pNativeContext != nullptr) {
     pNativeContext->refCount++;
     ret.reset(pNativeContext);
@@ -475,7 +475,7 @@ jobject WebPImage_nativeGetFrame(JNIEnv* pEnv, jobject thiz, jint index) {
   jobject ret = pEnv->NewObject(
       sClazzWebPFrame,
       sWebPFrameConstructor,
-      (jint) spFrameNativeContext.get());
+      (long) spFrameNativeContext.get());
   if (ret != nullptr) {
     // Ownership was transferred.
     spFrameNativeContext->refCount = 1;
@@ -526,7 +526,7 @@ std::unique_ptr<WebPFrameNativeContext, WebPFrameNativeContextReleaser>
   std::unique_ptr<WebPFrameNativeContext, WebPFrameNativeContextReleaser> ret(nullptr, releaser);
   pEnv->MonitorEnter(thiz);
   WebPFrameNativeContext* pNativeContext =
-      (WebPFrameNativeContext*) pEnv->GetIntField(thiz, sWebPFrameFieldNativeContext);
+      (WebPFrameNativeContext*) pEnv->GetLongField(thiz, sWebPFrameFieldNativeContext);
   if (pNativeContext != nullptr) {
     pNativeContext->refCount++;
     ret.reset(pNativeContext);
@@ -556,7 +556,7 @@ jint WebPImage_nativeGetSizeInBytes(JNIEnv* pEnv, jobject thiz) {
 void WebImage_nativeDispose(JNIEnv* pEnv, jobject thiz) {
   pEnv->MonitorEnter(thiz);
   WebPImageNativeContext* pNativeContext =
-      (WebPImageNativeContext*) pEnv->GetIntField(thiz, sWebPImageFieldNativeContext);
+      (WebPImageNativeContext*) pEnv->GetLongField(thiz, sWebPImageFieldNativeContext);
   if (pNativeContext != nullptr) {
     pEnv->SetIntField(thiz, sWebPImageFieldNativeContext, 0);
     WebPImageNativeContext_releaseRef(pEnv, thiz, pNativeContext);
@@ -761,7 +761,7 @@ jboolean WebPFrame_nativeShouldBlendWithPreviousFrame(JNIEnv* pEnv, jobject thiz
 void WebPFrame_nativeDispose(JNIEnv* pEnv, jobject thiz) {
   pEnv->MonitorEnter(thiz);
   WebPFrameNativeContext* pNativeContext =
-      (WebPFrameNativeContext*) pEnv->GetIntField(thiz, sWebPFrameFieldNativeContext);
+      (WebPFrameNativeContext*) pEnv->GetLongField(thiz, sWebPFrameFieldNativeContext);
   if (pNativeContext) {
     pEnv->SetIntField(thiz, sWebPFrameFieldNativeContext, 0);
     WebPFrameNativeContext_releaseRef(pEnv, thiz, pNativeContext);
